Add coverage_test module and fmt/lnt
